feat: add live clock component and integrate into wallet detail screen
- created a new component `LiveClock` with animated seconds display
- modified `texts.js` to add a new text entry for sharing
- integrated `LiveClock` into `WalletCardDetailScreen` with animation option

SVAK-184

# Conflicts:
#	src/screens/Wallet/WalletCardDetailScreen.tsx
