+++ src/components/index.js
@@ -75,6 +75,7 @@ export * from './Link';
 export * from './ListComponent';
 export * from './ListSettingsItem';
 export * from './ListSwitchItem';
+export * from './LiveClock';
 export * from './LiveTicker';
 export * from './LoadingContainer';
 export * from './LoadingModal';
